backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
pandas
ta
requests
python-dotenv
langgraph
langchain-core
groq
huggingface_hub
google-genai
feedparser
chromadb
sentence-transformers
redis
apscheduler
aiosqlite
pydantic
